requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
